refactor: move models and transformers
