+++ projects/movies/src/app/data-access/api/model/filter.interface.ts
@@ -1,0 +1,6 @@
+export interface TMDBFilterOptions {
+  dates?: {
+    maximum: string;
+    minimum: string;
+  };
+}
