+++ projects/movies/src/app/data-access/api/model/movie-category.model.ts
@@ -1,0 +1,5 @@
+import { TMDBMovieModel } from './movie.model';
+import { TMDBPaginatedResult } from './pagination.interface';
+import { TMDBFilterOptions } from './filter.interface';
+
+export type TMDBMovieCategoryModel = TMDBPaginatedResult<TMDBMovieModel> & TMDBFilterOptions
