+++ projects/movies/src/app/data-access/api/model/movie-person.model.ts
