+++ projects/movies/src/app/data-access/api/resources/discover.resource.ts
@@ -1,9 +1,9 @@
 import { TMDBMovieModel } from '../model/movie.model';
-import { baseUrlApiV3, getTMDBPaginationOptions, serverToClientPaginatedResult } from './utils';
-import { getHTTP } from '../../shared/injector/get-http-client';
+import { baseUrlApiV3, getTMDBPaginationOptions, serverToClientPaginatedResult } from '../utils';
+import { getHTTP } from '../../../shared/injector/get-http-client';
 import { TMDBPaginatedResult, TMDBPaginationOptions } from '../model/pagination.interface';
 import { map, Observable } from 'rxjs';
-import { PaginatedResult } from '../../shared/state/typings';
+import { PaginatedResult } from '../../../shared/state/typings';
 
 const URL_DISCOVER_MOVIE = [baseUrlApiV3, 'discover', 'movie'].join('/');
