+++ projects/movies/src/app/data-access/api/resources/genre.resource.ts
@@ -1,7 +1,7 @@
 import { map, Observable } from 'rxjs';
 import { TMDBMovieGenreModel } from '../model/movie-genre.model';
-import { baseUrlApiV3 } from './utils';
-import { getHTTP } from '../../shared/injector/get-http-client';
+import { baseUrlApiV3 } from '../utils';
+import { getHTTP } from '../../../shared/injector/get-http-client';
 
 const URL_GENRE_MOVIE_LIST = [baseUrlApiV3, 'genre', 'movie', 'list'].join('/');
 export const getGenres = (): Observable<TMDBMovieGenreModel[]> =>
