+++ projects/movies/src/app/data-access/api/resources/person.resource.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import { TMDBMoviePersonModel } from '../model/movie-person.model';
-import { baseUrlApiV3 } from './utils';
-import { getHTTP } from '../../shared/injector/get-http-client';
+import { baseUrlApiV3 } from '../utils';
+import { getHTTP } from '../../../shared/injector/get-http-client';
 
 const URL_PERSON = (id: string) => `${[baseUrlApiV3, 'person', id].join('/')}?append_to_response=videos`;
 export const getPerson = (id: string): Observable<TMDBMoviePersonModel> => {
