+++ projects/movies/src/app/data-access/api/resources/search.resource.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import { TMDBMovieModel } from '../model/movie.model';
-import { baseUrlApiV3 } from './utils';
-import { getHTTP } from '../../shared/injector/get-http-client';
+import { baseUrlApiV3 } from '../utils';
+import { getHTTP } from '../../../shared/injector/get-http-client';
 
 const URL_SEARCH = [baseUrlApiV3, 'search', 'movie'].join('/');
 export const getMoviesSearch = (
