--- projects/movies/src/app/data-access/model/movie-cast.model.ts
@@ -1,10 +1,0 @@
-export interface TMDBMovieCastModel {
-    cast_id: number;
-    character: string;
-    credit_id: string;
-    gender: number;
-    id: number;
-    name: string;
-    order: number;
-    profile_path: string;
-}
