--- projects/movies/src/app/data-access/model/movie-category.model.ts
@@ -1,12 +1,0 @@
-import {TMDBMovieModel} from './movie.model';
-
-export interface TMDBMovieCategoryModel {
-    page: number;
-    results: TMDBMovieModel[];
-    dates?: {
-        maximum: string;
-        minimum: string;
-    };
-    total_pages: number;
-    total_results: number;
-}
