--- projects/movies/src/app/data-access/model/movie-credits.model.ts
@@ -1,8 +1,0 @@
-import {TMDBMovieCastModel} from './movie-cast.model';
-import {TMDBMovieCrewModel} from './movie-crew.model';
-
-export interface TMDBMovieCreditsModel {
-    id: number;
-    cast: TMDBMovieCastModel[];
-    crew: TMDBMovieCrewModel[];
-}
