--- projects/movies/src/app/data-access/model/movie-crew.model.ts
@@ -1,9 +1,0 @@
-export interface TMDBMovieCrewModel {
-    credit_id: number;
-    department: string;
-    gender: number;
-    id: number;
-    job: string;
-    name: string;
-    profile_path: string;
-}
