--- projects/movies/src/app/data-access/model/movie-database.model.ts
@@ -1,14 +1,0 @@
-export interface TMDBMovieDatabaseModel {
-    id: number;
-    category: string;
-    date: string;
-    movieId: number;
-    original_title: string;
-    overview: string;
-    popularity: number;
-    poster_path: string;
-    release_date: string;
-    status: string;
-    userId: string;
-    watched: boolean;
-}
