+++ projects/movies/src/app/shared/state/discover.state.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { map } from 'rxjs';
-import { TMDBMovieModel } from '../../data-access/model/movie.model';
+import { TMDBMovieModel } from '../../data-access/api/model/movie.model';
 import { patch, RxState } from '@rx-angular/state';
 import { optimizedFetch } from '../utils/optimized-fetch';
 import { getActions } from '../rxa-custom/actions';
 import { withLoadingEmission } from '../utils/withLoadingEmissions';
-import { getDiscoverMovies } from '../../data-access/api/discover.resource';
+import { getDiscoverMovies } from '../../data-access/api/resources/discover.resource';
 
 export interface State {
   discoveredMovies: Record<string, TMDBMovieModel[]>;
