+++ projects/movies/src/app/shared/state/genre.state.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { exhaustMap } from 'rxjs';
-import { TMDBMovieGenreModel } from '../../data-access/model/movie-genre.model';
+import { TMDBMovieGenreModel } from '../../data-access/api/model/movie-genre.model';
 import { RxState } from '@rx-angular/state';
 import { getActions } from '../rxa-custom/actions';
-import { getGenres } from '../../data-access/api/genre.resource';
+import { getGenres } from '../../data-access/api/resources/genre.resource';
 
 export interface State {
   genres: TMDBMovieGenreModel[];
